[refactor] khushal87 Badge
