+++ src/Badge/withBadge.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
-import Badge, { BadgeProps } from './Badge';
+import Badge, { BadgeProps } from './index';
 
 type withBadgeOptions = {
   bottom?: number;
@@ -11,7 +11,7 @@ type withBadgeOptions = {
   containerStyle?: StyleProp<ViewStyle>;
 } & BadgeProps;
 
-const withBadge = (
+export const withBadge = (
   value: React.ReactNode | ((props: any) => React.ReactNode),
   options: withBadgeOptions = {}
 ) => (WrappedComponent: React.ComponentType<any>): React.ComponentType => {
@@ -63,5 +63,3 @@ const styles = StyleSheet.create({
     position: 'relative',
   },
 });
-
-export default withBadge;
