+++ src/index.ts
@@ -19,8 +19,7 @@ import SearchBar, { SearchBarProps } from './searchbar/SearchBar';
 import { SearchBarAndroidProps } from './searchbar/SearchBar-android';
 import { SearchBarDefaultProps } from './searchbar/SearchBar-default';
 import { SearchBarIosProps } from './searchbar/SearchBar-ios';
-import Badge, { BadgeProps } from './badge/Badge';
-import withBadge from './badge/withBadge';
+import Badge, { BadgeProps, withBadge } from './Badge';
 import CheckBox, { CheckBoxProps } from './Checkbox';
 import Divider, { DividerProps } from './Divider';
 import Slider, { SliderProps } from './Slider';
